+++ src/main/java/com/lovesoongalarm/lovesoongalarm/utils/JwtUtil.java
@@ -55,7 +55,7 @@ public class JwtUtil implements InitializingBean {
         return Jwts.builder()
                 .setHeaderParam(Header.JWT_TYPE, Header.JWT_TYPE)
                 .setClaims(claims)
-                .setIssuedAt(new Date(System.currentTimeMillis()))
+                .setIssuedAt(new Date(System.currentTimeMillis())) //현재 시간
                 .setExpiration(new Date(System.currentTimeMillis() + expiration))
                 .signWith(key)
                 .compact();
